Add colour sorting before image generation.
Adds both a "perceptive brightness" sort and a regular "hue" sort.
The hue sort results in a (subjectively) better image, but the sort
itself is quite slow (could improve by thinking hard to remove the
calls to atan2).

The performance of the tree structure is degraded significantly when
fed sorted colours.