Fix splitting plane bug that could leave leaf unsplit
If a leaf has values close to each other, the splitting plane might be
placed "at the edge" and end up leaving all pixels on one side due to
integer rounding errors.

The fix is to "shift" the plane along by one in the case the of integer
division resulting in truncation.