Rebuild tree periodically
Periodically (every 2% of the way through the pixel set) destroy and
recreate the spatial tree (using the wavefront of the resulting image).

This prevents the tree getting too "cluttered" with irrelevant nodes.

Gives a noticeable improvement to the running time with a "shuffled"
input (~10 minutes) and a *massive* improvement when given hue sorted
input (~100 minutes to ~3 minutes).