Split leaves by "maximum range" rather than "maximum balance"
Pick the splitting plane such that the colour with the largest range
is chosen to be split on, rather than the plane that would lead to the
most balanced children.

In danger of over-optimising for the hue-sorted case, but this reduces
the running time to around a minute (from three).