Fix bug distance comparison bug in tree search
The distance between the query point and the splitting plane was
being compared to the *square* of the distance between the query
point and the "best" match from the child node.

Since all distances are usually "squared distances" (to avoid
calculating square roots), the solution is simply to square the
distance to the splitting plane too.

This bug was the cause of the poor performance for most "initial
orderings" - the "hue sorted" case is not massively affected, but the
other cases have their running times brought down to a minute or two
(from ~10-15 minutes).