Trigger tree rebuild on number of empty leaves
Rather than rebuild every "2%" of the way through the pixels to place,
use a heuristic to decide when the tree is "unsuitable".

It's quite difficult to determine exactly what causes the performance
of the tree to degrade, but a large number of empty leaves does appear
to cause problems.

Rebuilding when the number of empty leaves reaches 5% seems reasonable.